Enable `robustness` transform by default in `tint`.
Update the `tint` command to always enable `robustness`. The
`robustness` option is removed from the list of possible transforms
which can be requested. A new `--disable-robustness` flag is added in
order to generate code with robustness disabled.

Bug: 378726537
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216996
Commit-Queue: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>

+++ test/tint/buffer/uniform/std140/struct/mat2x2_f32/to_storage.wgsl.expected.glsl
@@ -119,8 +119,8 @@ void main() {
     }
   }
   tint_store_and_preserve_padding(v_5);
-  S v_8 = tint_convert_S(v.inner[2]);
-  tint_store_and_preserve_padding_1(uint[1](uint(1)), v_8);
-  v_1.inner[3].m = mat2(v.inner[2].m_col0, v.inner[2].m_col1);
-  v_1.inner[1].m[0] = v.inner[0].m_col1.yx;
+  S v_8 = tint_convert_S(v.inner[2u]);
+  tint_store_and_preserve_padding_1(uint[1](1u), v_8);
+  v_1.inner[3u].m = mat2(v.inner[2u].m_col0, v.inner[2u].m_col1);
+  v_1.inner[1u].m[0u] = v.inner[0u].m_col1.yx;
 }
