+++ test/tint/builtins/workgroupUniformLoad/via_param.wgsl.expected.glsl
@@ -8,7 +8,7 @@ int foo(uint p_indices[1]) {
   return v_1;
 }
 int bar() {
-  return foo(uint[1](uint(0)));
+  return foo(uint[1](0u));
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
