+++ test/tint/ptr_ref/load/param/storage/vec2_f32_in_mat2x2.wgsl.expected.glsl
@@ -9,5 +9,5 @@ vec2 func(uint pointer_indices[1]) {
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
-  vec2 r = func(uint[1](uint(1)));
+  vec2 r = func(uint[1](1u));
 }
