+++ test/tint/ptr_ref/load/param/uniform/struct_in_array.wgsl.expected.glsl
@@ -14,5 +14,5 @@ str func(uint pointer_indices[1]) {
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
-  str r = func(uint[1](uint(2)));
+  str r = func(uint[1](2u));
 }
