+++ test/tint/ptr_ref/store/param/workgroup/struct_in_array.wgsl.expected.glsl
@@ -26,7 +26,7 @@ void tint_symbol_inner(uint tint_local_index) {
     }
   }
   barrier();
-  func(uint[1](uint(2)));
+  func(uint[1](2u));
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
